common layout and fixed links
